refactor(Evaluation): extract to seperate file
